my_script.py
import pandas as pd   
import numpy as np
import lime
import lime.lime_tabular
import pickle

import lightgbm as lgb

import matplotlib.pyplot as plt
import plotly.express as px
import seaborn as sns

from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_curve, roc_auc_score
from sklearn.metrics import accuracy_score