requirements.txt
lime==0.2.0.1
lightgbm==3.2.1
scikit-learn==0.24.2
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
plotly-express==0.5.0
seaborn==0.11.1